// src/app/shared/models/history.model.ts
export class HistoryModel {
  created: any;
  lastUpdated: any;
  player1Username: string;
  player1Name: string;
  player1Uid: string;
  player1NewRating: number;
  player1OldRating: number;
  player2NewRating: number;
  player2OldRating: number;
  player2Uid: string;
  player2Username: string;
  player2Name: string;
  player1Won: boolean;

}